# Reuse and optimize sorting logic for `List`, `SelfList`, and `HashMap`