Added `SortList` class, and updated `List`, `SelfList`, and `HashMap` sort methods to use it.  Sorting is done with merge sort, with an initial check to optimize for already sorted lists, and sorted lists that were appended to.

Removed `List::sort_custom_inplace`, since `sort_custom` no longer allocates additional memory.

Added more sorting tests.

Speeds up `ClassDB.get_class_list()` by around 10% when called from GDScript, and speeds up `Dictionary.sort()`, especially for larger unsorted dictionaries, as worst case time complexity is now O(n\*logn) instead of O(n\*n).  Compared with GDScript below:

```gdscript
var shuffled_dictionaries : Array[Dictionary] = []

func _ready() -> void:
	run_test("get_class_list")
	run_test("sorted_dictionary")
	run_test("nearly_sorted_dictionary")
	generate_shuffled_dictionaries(25, 100000)
	run_test("small_shuffled_dictionaries")
	run_test("large_shuffled_dictionary")
	
func run_test(p_name : String):
	var start := Time.get_ticks_msec()
	call(p_name)
	var end := Time.get_ticks_msec()
	print("%s: %dms" % [p_name, end - start])

func get_class_list():
	for i in 1000:
		ClassDB.get_class_list()

func sorted_dictionary():
	var dictionary := {}
	var keys := range(25).map(str)
	keys.sort()
	for key in keys:
		dictionary[key] = true
	
	for i in 1000000:
		dictionary.sort()

func nearly_sorted_dictionary():
	var dictionary := {}
	for i in 25:
		dictionary[i] = true;
	for i in 500000:
		i = i % 25
		dictionary.erase(i)
		dictionary[i] = true;
		dictionary.sort()

func generate_shuffled_dictionaries(p_size: int, p_count: int):
	var keys := range(p_size).map(str)
	for i in p_count:
		keys.shuffle()
		var d := {}
		for j in p_size:
			d[keys[j]] = true
		shuffled_dictionaries.push_back(d)

func small_shuffled_dictionaries():
	for d in shuffled_dictionaries:
		d.sort()

func large_shuffled_dictionary():
	var keys = range(10000)
	keys.shuffle()
	var d := {}
	for key in keys:
		d[key] = true
	d.sort()
```

old
```
get_class_list: 277ms
sorted_dictionary: 595ms
nearly_sorted_dictionary: 468ms
small_shuffled_dictionaries: 434ms
large_shuffled_dictionary: 399ms
```

new
```
get_class_list: 254ms
sorted_dictionary: 575ms
nearly_sorted_dictionary: 456ms
small_shuffled_dictionaries: 342ms
large_shuffled_dictionary: 6ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->