@aaronp64 2025-04-22
I took a look at the failed check, there was a warning in the Setup Vulkan SDK section:
```
curl: (56) Recv failure: Connection reset by peer
Warning: macOS: Vulkan SDK installation failed, building without Vulkan support.
```
The next section looks like it failed due to a Vulkan related macro not being defined.  I think this should be ok if the check can just be rerun, but let me know if I need to update anything for this.